Add support for archiving blob
Method `IAzureStore.ArchiveBlobAsync()` can be used to copy a 
blob to an "archive" container in the same storage account, compress it
(using GZipStream) and change its storage tier to "Archive". The original
blob remains in the "persist" container, and will need to be deleted
in another way.

Method `IAzureStore.TryUnArchiveBlobAsync()` can be used to 
copy the blob from the "archive" container to the "staging" container,
change its storage tier to "Hot", and then uncompress and write 
its contents back to "persist". Note that the rehydration stage
can take quite a long time, so it may be necessary to call this 
function repeatedly until it returns a success status.